chore(tests): move ViewRenderingTestCase to integration tests
ref #13860
